🔵 extract Cron()
